doc_chat_app/backend/requirements.txt
fastapi
uvicorn
python-docx
langchain
chromadb
python-multipart
tiktoken
beautifulsoup4
lxml
sentence-transformers
huggingface-hub
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-core
torch
unstructured
langchain-community
